revert thanm to old version
40481e20ce4472f094931abc68c654077f42e670
aka the merge base between master and thanm-new-spec-format

+++ thanm/image.c
@@ -42,7 +42,7 @@ unsigned int
 format_Bpp(
     format_t format)
 {
-    switch ((int)format) {
+    switch (format) {
     case FORMAT_RGBA8888:
     case FORMAT_BGRA8888:
         return 4;
